2022 11 16

+++ lib/clean_architecture/domain/model/order.dart
@@ -565,12 +565,12 @@ class VariationPrice {
     this.sale,
   });
 
-  final int? regular;
-  final int? sale;
+  final double? regular;
+  final double? sale;
 
   VariationPrice copyWith({
-    int? regular,
-    int? sale,
+    double? regular,
+    double? sale,
   }) =>
       VariationPrice(
         regular: regular ?? this.regular,
@@ -578,8 +578,8 @@ class VariationPrice {
       );
 
   factory VariationPrice.fromMap(Map<String, dynamic> json) => VariationPrice(
-        regular: json["regular"] == null ? null : json["regular"],
-        sale: json["sale"] == null ? null : json["sale"],
+        regular: json["regular"] == null ? null : json["regular"].toDouble(),
+        sale: json["sale"] == null ? null : json["sale"].toDouble(),
       );
 
   Map<String, dynamic> toMap() => {
@@ -624,12 +624,12 @@ class Regular {
     this.max,
   });
 
-  final int? min;
-  final int? max;
+  final double? min;
+  final double? max;
 
   Regular copyWith({
-    int? min,
-    int? max,
+    double? min,
+    double? max,
   }) =>
       Regular(
         min: min ?? this.min,
